test:
- added test code demonstrating makePauseInvitation working as expected
- removed instances of unnecessary duplicate tests assertions
- renamed e2e.test.js to coreEval.test.js
